events: Add Event class
The Event class is a multi-handler dispatcher with flexible listener
registration and deregistration (called attach() and detach()).

While modular project designs will create a file for each specific used
Event, it's also intended for inline use, since most times you only
need to derive onRaise(). A more robust implementation would also use
isInvocationValid(), but this is most times a one-liner and is also
suited for inlining in type-declaration.

This commit also introduces the TestUtilities needed for the EventTest.